fix:  ❌refactored docs store and fixed map|filter errors in getters, successfully setup tests
