+++ src/store/__tests__/docs.test.js
@@ -1,16 +1,26 @@
 import { createLocalVue } from '@vue/test-utils';
 import Vuex from 'vuex';
-import * as docs from '../docs';
+import store from '..';
+import { types as mutations } from '../docs';
+
+import { resetState, DEFAULT_STATE } from '../helpers/resetState';
 
 const localVue = createLocalVue();
 localVue.use(Vuex);
 
 describe('Test for docs operations', () => {
-  const store = new Vuex.Store({
-    modules: {
-      docs
-    }
+  beforeEach(resetState);
+
+  test('Clean state', () => {
+    console.log({ DEFAULT_STATE });
+    expect(store.state).toStrictEqual(DEFAULT_STATE);
   });
 
-  console.log(store.state.actions.cwd);
+  // console.log(JSON.stringify(store.state, null, 2));
+  // store.commit(mutations.SET_CWD, '/home/divine/Desktop');
+  // console.log(JSON.stringify(store.state, null, 2));
+
+  test('truthy', () => {
+    expect(true).toBe(true);
+  });
 });
